backtesting/requirements.txt
# Required Python packages for the backtesting system

clickhouse-connect>=0.7.0 # For ClickHouse interaction
pandas>=1.3.0        # For data manipulation
python-dotenv>=0.19.0 # For loading .env files
scikit-learn>=1.0.0  # For loading the pickled model (adjust version if needed based on your model)
openpyxl>=3.0.0      # Potentially needed by pandas for Excel export/read (optional but good to have)
requests>=2.25.0     # Potentially used by underlying libraries
# Add any other specific dependencies your model might require

# Added pytz for timezone handling in db_utils
pytz>=2021.3 